More tests that need the core generators.
